requirements.txt
python-dotenv==0.21.1
python-docx==0.8.11
pyTelegramBotAPI==4.9.0
